Move client to top level
