+++ crates/bitwarden-user-crypto-management/src/client.rs
